Improves resource deletion handling
Refactors the resource deletion process to enhance error handling.
It removes unnecessary logic from the controller and moves the validation to the query builder
so if a resource does not exist, it throws an exception, instead of returning null.

Updates the response to return json objects.
Adds authorization header to request.